refactor (step3) 단순 패키지 분리
