+++ src/main/java/kitchenpos/ordertable/controller/TableGroupRestController.java
@@ -1,6 +1,6 @@
 package kitchenpos.controller;
 
-import kitchenpos.bo.TableGroupBo;
+import kitchenpos.ordertable.bo.TableGroupBo;
 import kitchenpos.model.TableGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
