+++ src/main/java/kitchenpos/ordertable/controller/TableRestController.java
@@ -1,6 +1,6 @@
 package kitchenpos.controller;
 
-import kitchenpos.bo.TableBo;
+import kitchenpos.ordertable.bo.TableBo;
 import kitchenpos.model.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
